+++ src/ui/messages/layer-titled-message-view-container.js
@@ -66,7 +66,7 @@ registerComponent('layer-titled-message-view-container', {
     },
 
     /**
-     * Title for the titlebar; comes from `this.properties.ui._getTitle()`
+     * Title for the titlebar; comes from `this.properties.ui.getTitle()`
      *
      * @property {String} title
      */
@@ -78,7 +78,7 @@ registerComponent('layer-titled-message-view-container', {
     },
 
     /**
-     * Icon for the titlebar; comes from `this.properties.ui._getIconClass()`
+     * Icon for the titlebar; comes from `this.properties.ui.getIconClass()`
      *
      * @property {String} icon
      */
@@ -97,8 +97,8 @@ registerComponent('layer-titled-message-view-container', {
     },
 
     onRerender() {
-      this.icon = this.properties.ui._getIconClass();
-      this.title = this.properties.ui._getTitle();
+      this.icon = this.properties.ui.getIconClass();
+      this.title = this.properties.ui.getTitle();
     },
   },
 });